TRUS-394 Postcode for trust is Option[String]
Postcode for trust now takes in an optional string, which if nonEmpty
constrains to a postcode regex and validates.
If not empty and a malformed postcode it errors.
If form is empty it validates successfully and continues.

Implemented a OptionalStringPage which ensures a value is in the field
if defined, otherwise checks it's empty.